refactor: media moved AuthenticatedMedia in its own file used useAvatarUrl (renamed to useMediaUrl) to fetch Media
