+++ src/components/textual-room/message-item.tsx
@@ -1,12 +1,12 @@
 import { useMatrixClient } from "@/hooks/use-matrix-client";
-import { Pen, Trash, FileIcon, Loader2, Download, X } from "lucide-react";
+import { Pen, Trash } from "lucide-react";
 import { MatrixEventEvent, MsgType, RelationType, EventType } from "matrix-js-sdk";
 import type { MatrixEvent } from "matrix-js-sdk";
 import type { FC } from "react";
 import type React from "react";
 import { useEffect, useState } from "react";
+import { AuthenticatedMedia } from "@/components/textual-room/authenticated-media";
 import { Button } from "../ui/button";
-import { Dialog, DialogContent, DialogTitle, DialogTrigger, DialogClose } from "../ui/dialog";
 import { Textarea } from "../ui/textarea";
 
 interface ReactionContent {
@@ -53,204 +53,6 @@ const collectReactionsByEmoji = (events: MatrixEvent[], targetEventId: string) =
     }
 
     return byEmoji;
-};
-
-const formatFileSize = (bytes?: number) => {
-    if (bytes == null) return "Taille inconnue";
-    const mb = bytes / (1024 * 1024);
-    if (mb < 1) return `${(bytes / 1024).toFixed(1)} Ko`;
-    return `${mb.toFixed(2)} Mo`;
-};
-
-const AuthenticatedMedia: FC<{
-    mxcUrl: string;
-    msgtype: string;
-    body: string;
-    fileSize?: number;
-    onDialogOpenChange?: (open: boolean) => void;
-}> = ({ mxcUrl, msgtype, body, fileSize, onDialogOpenChange }) => {
-    const { client } = useMatrixClient();
-    const [objectUrl, setObjectUrl] = useState<string | null>(null);
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        setLoading(true);
-        setObjectUrl(null);
-
-        let url: string | null = null;
-        const controller = new AbortController();
-
-        const fetchMedia = async () => {
-            const accessToken = client.getAccessToken();
-
-            const httpUrl = client.mxcUrlToHttp(
-                mxcUrl,
-                undefined,
-                undefined,
-                undefined,
-                true,
-                true,
-                true
-            );
-
-            if (!accessToken || !httpUrl) {
-                if (!controller.signal.aborted) setLoading(false);
-                return;
-            }
-
-            try {
-                const res = await fetch(httpUrl, {
-                    signal: controller.signal,
-                    headers: { Authorization: `Bearer ${accessToken}` }
-                });
-
-                if (!res.ok) {
-                    console.error(`Media fetch error: ${res.status.toString()} ${res.statusText}`);
-                    if (!controller.signal.aborted) setLoading(false);
-                    return;
-                }
-
-                const blob = await res.blob();
-                if (!controller.signal.aborted) {
-                    url = URL.createObjectURL(blob);
-                    setObjectUrl(url);
-                }
-            } catch (error) {
-                if (error instanceof Error && error.name === "AbortError") return;
-                console.error("Media fetch error:", error);
-            } finally {
-                if (!controller.signal.aborted) setLoading(false);
-            }
-        };
-
-        void fetchMedia();
-
-        return () => {
-            controller.abort();
-            if (url) URL.revokeObjectURL(url);
-        };
-    }, [client, mxcUrl]);
-
-    if (loading) {
-        return (
-            <div className="mt-2 flex w-fit items-center gap-2 rounded-md border p-3 text-sm text-slate-500">
-                <Loader2 className="size-4 animate-spin" /> Chargement...
-            </div>
-        );
-    }
-
-    if (!objectUrl) {
-        return <div className="mt-2 text-sm text-red-500">Erreur de chargement du média</div>;
-    }
-
-    if (msgtype === "m.image") {
-        return (
-            <Dialog onOpenChange={onDialogOpenChange}>
-                <DialogTrigger
-                    render={
-                        <button className="focus-visible:ring-ring mt-2 max-w-sm cursor-pointer overflow-hidden rounded-md border bg-gray-50/50 transition-opacity outline-none hover:opacity-90 focus-visible:ring-2 focus-visible:ring-offset-2" />
-                    }
-                >
-                    <img src={objectUrl} alt={body} className="max-h-96 w-full object-contain" />
-                </DialogTrigger>
-
-                <DialogContent
-                    showCloseButton={false}
-                    className="top-0! left-0! flex h-dvh w-dvw max-w-none! translate-x-0! translate-y-0! items-center justify-center border-none bg-transparent p-0! shadow-none ring-0"
-                >
-                    <DialogTitle className="sr-only">Aperçu de l&#39;image : {body}</DialogTitle>
-
-                    <DialogClose
-                        className="absolute inset-0 z-0 h-full w-full cursor-default outline-none"
-                        aria-hidden="true"
-                        tabIndex={-1}
-                    />
-
-                    <img
-                        src={objectUrl}
-                        alt={body}
-                        className="relative z-10 max-h-[90vh] max-w-[90vw] rounded-md object-contain"
-                    />
-
-                    <DialogClose className="bg-background text-foreground hover:bg-muted focus:ring-ring absolute top-4 right-4 z-50 flex size-8 items-center justify-center rounded-md border shadow-sm transition-colors focus:ring-2 focus:outline-none">
-                        <X className="size-4" />
-                        <span className="sr-only">Fermer</span>
-                    </DialogClose>
-                </DialogContent>
-            </Dialog>
-        );
-    }
-
-    if (msgtype === "m.video") {
-        return (
-            <div className="group relative mt-2 w-fit overflow-hidden rounded-md border bg-transparent shadow-sm">
-                <video
-                    src={objectUrl}
-                    controls
-                    preload="metadata"
-                    className="max-h-96 w-auto max-w-sm object-contain"
-                />
-
-                <a
-                    href={objectUrl}
-                    download={body}
-                    title="Télécharger"
-                    aria-label="Télécharger la vidéo"
-                    className="bg-background text-foreground hover:bg-muted absolute top-2 right-2 flex size-8 items-center justify-center rounded-md border opacity-0 shadow-sm transition-all duration-200 group-hover:opacity-100 hover:scale-105"
-                >
-                    <Download className="size-4" />
-                </a>
-            </div>
-        );
-    }
-
-    if (msgtype === "m.audio") {
-        return (
-            <div className="group relative mt-2 w-fit rounded-xl border bg-gray-50/50 p-2 shadow-sm">
-                <div className="text-muted-foreground mb-1.5 flex items-center justify-between px-2 text-xs">
-                    <span
-                        className="max-w-40 truncate font-medium text-slate-700 md:max-w-50"
-                        title={body}
-                    >
-                        {body}
-                    </span>
-                    <span className="shrink-0 text-[10px]">{formatFileSize(fileSize)}</span>
-                </div>
-
-                <div className="overflow-hidden rounded-full border bg-white shadow-sm">
-                    <audio
-                        src={objectUrl}
-                        controls
-                        preload="metadata"
-                        className="h-10 w-64 md:w-72"
-                    />
-                </div>
-
-                <a
-                    href={objectUrl}
-                    download={body}
-                    title="Télécharger"
-                    aria-label={`Télécharger ${body}`}
-                    className="bg-background text-foreground hover:bg-muted absolute -top-2 -right-2 flex size-8 items-center justify-center rounded-md border opacity-0 shadow-sm transition-all duration-200 group-hover:opacity-100 hover:scale-105"
-                >
-                    <Download className="size-4" />
-                </a>
-            </div>
-        );
-    }
-
-    return (
-        <a
-            href={objectUrl}
-            download={body}
-            className="mt-2 flex w-fit items-center gap-2 rounded-md border bg-gray-50 p-3 transition-colors hover:bg-gray-100"
-        >
-            <FileIcon className="size-5 text-indigo-500" />
-            <span className="text-sm font-medium text-slate-700 underline-offset-4 hover:underline">
-                {body}
-            </span>
-        </a>
-    );
 };
 
 const MessageItem: FC<{ event: MatrixEvent; roomId: string }> = ({ event, roomId }) => {
